+++ packages/storybook/src/components/highlight-box/highlight-box.core-template.ts
@@ -1,5 +1,5 @@
 import { HighlightBox } from "@dso-toolkit/sources";
-import { html } from "lit-html";
+import { html, nothing } from "lit-html";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 import { unsafeHTML } from "lit-html/directives/unsafe-html.js";
 import { ComponentImplementation } from "../../templates";
@@ -7,8 +7,8 @@ import { ComponentImplementation } from "../../templates";
 export const coreHighlightBox: ComponentImplementation<HighlightBox> = {
   component: "highlightBox",
   implementation: "core",
-  template: ({ iconTemplate }) =>
-    function highlightBoxTemplate({ yellow, white, dropShadow, border, step, icon, richContent }) {
+  template: ({ iconTemplate, imageTemplate }) =>
+    function highlightBoxTemplate({ yellow, white, dropShadow, border, step, icon, richContent, bannerImage }) {
       return html`
         <dso-highlight-box
           step=${ifDefined(typeof step === "number" && step > 0 ? step : undefined)}
@@ -18,7 +18,14 @@ export const coreHighlightBox: ComponentImplementation<HighlightBox> = {
           ?border=${border}
         >
           ${icon && iconTemplate({ icon, slot: "icon" })}
-          <div class="dso-rich-content">${typeof richContent === "string" ? unsafeHTML(richContent) : richContent}</div>
+          ${bannerImage
+            ? imageTemplate({
+                source: "images/man-vrouw-kaart.png",
+                alt: "man en vrouw met kaart",
+                modifier: "dso-highlight-box-banner",
+              })
+            : nothing}
+          ${typeof richContent === "string" ? unsafeHTML(richContent) : richContent}
         </dso-highlight-box>
       `;
     },